working

+++ salre/src/main/java/com/salre/main/notification/EmitterRepository.java
@@ -12,7 +12,7 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class EmitterRepository {
     // 모든 Emitters를 저장하는 ConcurrentHashMap
-    private final Map<Long, SseEmitter> emitters = new ConcurrentHashMap<>();
+    private final Map<Integer, SseEmitter> emitters = new ConcurrentHashMap<>();
 
     /**
      * 주어진 아이디와 이미터를 저장
@@ -20,7 +20,7 @@ public class EmitterRepository {
      * @param id      - 사용자 아이디.
      * @param emitter - 이벤트 Emitter.
      */
-    public void save(Long id, SseEmitter emitter) {
+    public void save(int id, SseEmitter emitter) {
         emitters.put(id, emitter);
     }
 
@@ -29,7 +29,7 @@ public class EmitterRepository {
      *
      * @param id - 사용자 아이디.
      */
-    public void deleteById(Long id) {
+    public void deleteById(int id) {
         emitters.remove(id);
     }
 
@@ -39,7 +39,7 @@ public class EmitterRepository {
      * @param id - 사용자 아이디.
      * @return SseEmitter - 이벤트 Emitter.
      */
-    public SseEmitter get(Long id) {
+    public SseEmitter get(int id) {
         return emitters.get(id);
     }
 }
